Move JunctionBuffer to within the method
